Support incoming relations
Use the Timbuctoo proprety `name` as id for data extension.
The property 'name' is more unique than the 'uri' in Timbuctoo.
The 'uri' is the same for incoming and outgoing relations.
The 'name' has a prefix '_inverse' for incoming relations.
For outgoing relation the 'name' will not have a prefix.